environment-api/requirements.txt
fastapi==0.86.0
uvicorn[standard]==0.19.0
httpx==0.23.1
starlette==0.20.4
python-multipart==0.0.5
numpy==1.24.2
opencv-python==4.7.0.72
fastapi-utils==0.2.1
boto3~=1.26.130
Pillow~=9.5.0
torch==1.13.1
scikit-image~=0.21.0
minecraft-data==3.20.0